rename: main 폴더 home으로 변경
