+++ src/pages/home/Home.tsx
@@ -1,7 +1,15 @@
-import React from 'react';
+import FilterList from '../home/components/filterList/FilterList';
+import SearchBar from '../home/components/searchBar/SearchBar';
+import PostList from '../home/components/postLIst/PostList';
 
 function Home() {
-  return <div>Home</div>;
+  return (
+    <section>
+      <FilterList />
+      <SearchBar />
+      <PostList />
+    </section>
+  );
 }
 
 export default Home;
